others/simPhi/requirements.txt
eyed3
tinytag
Pillow
matplotlib